app/requirements.txt
flask
wand
pillow
keyboard
brother_ql_next == 0.11.3
waveshare-epaper
gpiozero
lgpio
